[Feature] Introduce job log
