--- paimon-web-server/src/main/java/org/apache/paimon/web/server/context/logTool/LogType.java
@@ -1,50 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.paimon.web.server.context.logTool;
-
-/** support batch streaming */
-public enum LogType {
-    STREAMING("Streaming"),
-
-    BATCH("Batch"),
-    UNKNOWN("Unknown");
-
-    private String value;
-
-    LogType(String value) {
-        this.value = value;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    public static LogType get(String value) {
-        for (LogType type : LogType.values()) {
-            if (type.getValue().equalsIgnoreCase(value)) {
-                return type;
-            }
-        }
-        return LogType.UNKNOWN;
-    }
-
-    public boolean equalsValue(String type) {
-        return value.equalsIgnoreCase(type);
-    }
-}
