+++ paimon-web-server/src/main/java/org/apache/paimon/web/server/service/impl/JobServiceImpl.java
@@ -28,9 +28,9 @@ import org.apache.paimon.web.gateway.config.ExecutionConfig;
 import org.apache.paimon.web.gateway.enums.EngineType;
 import org.apache.paimon.web.gateway.provider.ExecutorFactoryProvider;
 import org.apache.paimon.web.server.context.LogContextHolder;
-import org.apache.paimon.web.server.context.logTool.LogEntity;
-import org.apache.paimon.web.server.context.logTool.LogReadPool;
-import org.apache.paimon.web.server.context.logTool.LogType;
+import org.apache.paimon.web.server.context.logtool.LogEntity;
+import org.apache.paimon.web.server.context.logtool.LogReadPool;
+import org.apache.paimon.web.server.context.logtool.LogType;
 import org.apache.paimon.web.server.data.dto.JobSubmitDTO;
 import org.apache.paimon.web.server.data.dto.ResultFetchDTO;
 import org.apache.paimon.web.server.data.dto.SessionDTO;
